chore: move openrouter error check test to models test dir
