+++ tests/models/cassettes/test_openrouter/test_openrouter_errors_raised.yaml
